@@ -7,8 +7,8 @@ import 'package:morea/services/utilities/morea_functions.dart';
 import 'package:morea/services/utilities/qr_code.dart';
 import 'package:rxdart/rxdart.dart';
 
-Future<Widget> makeLeiterWidget(
-    BuildContext context, String userID, String groupID) {
+Future<Widget> makeLeiterWidget(BuildContext context, String userID,
+    String groupID) {
   QrCode qrCode = new QrCode();
   StreamController stream = new BehaviorSubject();
   bool buttonPressed = false;
@@ -33,7 +33,7 @@ Future<Widget> makeLeiterWidget(
         if (!buttonPressed)
           return Column(
             children: <Widget>[
-              new RaisedButton(
+              new ElevatedButton(
                 child: Row(
                   mainAxisSize: MainAxisSize.min,
                   children: <Widget>[
@@ -44,35 +44,46 @@ Future<Widget> makeLeiterWidget(
                     new Text('Scannen', style: new TextStyle(fontSize: 20))
                   ],
                 ),
-                onPressed: () async => {
-                  buttonPressed = true,
-                  await qrCode.germanScanQR(),
-                  stream.add(666),
+                onPressed: () async {
+                  buttonPressed = true;
+                  await qrCode.germanScanQR();
+                  stream.add(666);
+                  print(qrCode.germanError);
                   if (qrCode.germanError ==
-                      'Um den Kopplungsvorgang mit dem TN abzuschliessen, scanne den Qr-Code, der im Profil des TN\'s ersichtlich ist.')
-                    {
-                      makeLeiter(userID, qrCode.qrResult, groupID)
-                          .then((onValue) {
-                        stream.add("nicht dicht!");
-                      })
-                    }
+                      'Um den Kopplungsvorgang mit deinem Kind abzuschliessen, scanne den Qr-Code, der im Profil deines Kindes ersichtlich ist.') {
+                    print('making Leiter');
+                    makeLeiter(userID, qrCode.qrResult, groupID)
+                        .then((onValue) {
+                      stream.add("nicht dicht!");
+                    });
+                  }
                   else
-                    buttonPressed = false
+                    buttonPressed = false;
                 },
-                shape: new RoundedRectangleBorder(
-                    borderRadius: new BorderRadius.circular(30.0)),
-                color: Color(0xff7a62ff),
-                textColor: Colors.white,
+                style: ButtonStyle(
+                  shape: MaterialStateProperty.all<OutlinedBorder>(
+                      RoundedRectangleBorder(
+                          borderRadius: BorderRadius.circular(30))),
+                  backgroundColor:
+                  MaterialStateProperty.all<Color>(Color(0xff7a62ff)),
+                  foregroundColor:
+                  MaterialStateProperty.all<Color>(Colors.white),
+                ),
               ),
-              new RaisedButton(
+              ElevatedButton(
                 child:
-                    new Text('Abbrechen', style: new TextStyle(fontSize: 20)),
+                new Text('Abbrechen', style: new TextStyle(fontSize: 20)),
                 onPressed: () async =>
-                    {stream.close(), Navigator.of(context).pop()},
-                shape: new RoundedRectangleBorder(
-                    borderRadius: new BorderRadius.circular(30.0)),
-                color: Color(0xff7a62ff),
-                textColor: Colors.white,
+                {stream.close(), Navigator.of(context).pop()},
+                style: ButtonStyle(
+                  shape: MaterialStateProperty.all<OutlinedBorder>(
+                      RoundedRectangleBorder(
+                          borderRadius: BorderRadius.circular(30))),
+                  backgroundColor:
+                  MaterialStateProperty.all<Color>(Color(0xff7a62ff)),
+                  foregroundColor:
+                  MaterialStateProperty.all<Color>(Colors.white),
+                ),
               ),
             ],
           );
@@ -95,17 +106,23 @@ Future<Widget> makeLeiterWidget(
               SizedBox(
                 height: 200,
               ),
-              new RaisedButton(
+              ElevatedButton(
                 child:
-                    new Text('Abbrechen', style: new TextStyle(fontSize: 20)),
+                new Text('Abbrechen', style: new TextStyle(fontSize: 20)),
                 onPressed: () async =>
-                    {stream.close(), Navigator.of(context).pop()},
-                shape: new RoundedRectangleBorder(
-                    borderRadius: new BorderRadius.circular(30.0)),
-                color: Color(0xff7a62ff),
-                textColor: Colors.white,
+                {stream.close(), Navigator.of(context).pop()},
+                style: ButtonStyle(
+                  shape: MaterialStateProperty.all<OutlinedBorder>(
+                      RoundedRectangleBorder(
+                          borderRadius: BorderRadius.circular(30))),
+                  backgroundColor:
+                  MaterialStateProperty.all<Color>(Color(0xff7a62ff)),
+                  foregroundColor:
+                  MaterialStateProperty.all<Color>(Colors.white),
+                ),
               ),
-            ],
+            ]
+            ,
           );
       },
     ),
