@@ -1,15 +1,17 @@
 import 'package:flutter/material.dart';
 import 'package:morea/Widgets/standart/moreaTextStyle.dart';
 import 'package:morea/morea_strings.dart';
+import 'package:morea/services/crud.dart';
 import 'package:morea/services/morea_firestore.dart';
 import 'package:morea/services/auth.dart';
 import 'package:morea/morealayout.dart';
 
 class SendMessages extends StatefulWidget {
-  SendMessages({this.moreaFire, this.auth});
+  SendMessages({this.moreaFire, this.auth, this.crudMedthods});
 
   final MoreaFirebase moreaFire;
   final Auth auth;
+  final CrudMedthods crudMedthods;
 
   @override
   State<StatefulWidget> createState() {
@@ -19,6 +21,7 @@ class SendMessages extends StatefulWidget {
 
 class _SendMessagesState extends State<SendMessages> {
   MoreaFirebase moreaFire;
+  CrudMedthods crudMedthods;
   String uid;
   bool loading = true;
   GlobalKey<FormState> _formKey = GlobalKey<FormState>();
@@ -30,20 +33,38 @@ class _SendMessagesState extends State<SendMessages> {
   FocusNode inhaltFocus = FocusNode();
   TextEditingController vorschauController = TextEditingController();
   FocusNode vorschauFocus = FocusNode();
+  List<Map<String, dynamic>> subgroups = <Map<String, dynamic>>[];
+  bool initDone = false;
 
-  bool biberCheckBox = false;
-  bool woelfeCheckBox = false;
-  bool meitliCheckBox = false;
-  bool buebeCheckBox = false;
+  Map<String, bool> groupCheckbox;
 
   @override
   void initState() {
     super.initState();
     moreaFire = widget.moreaFire;
+    this.crudMedthods = widget.crudMedthods;
     getUID();
+    initSubgroups();
   }
 
-  void getUID() async{
+  void initSubgroups() async {
+    Map<String, dynamic> data =
+        (await this.crudMedthods.getDocument(pathGroups, moreaGroupID)).data();
+    data[groupMapGroupOption][groupMapGroupLowerClass]
+        .forEach((key, value) => this.subgroups.add(value));
+    this.groupCheckboxinit(this.subgroups);
+    this.initDone = true;
+    setState(() {});
+  }
+
+  void groupCheckboxinit(List<Map<String, dynamic>> subgroups) {
+    this.groupCheckbox = Map<String, bool>();
+    for (Map<String, dynamic> groupMap in subgroups) {
+      this.groupCheckbox[groupMap['groupID']] = false;
+    }
+  }
+
+  void getUID() async {
     uid = await widget.auth.currentUser();
   }
 
@@ -70,240 +91,194 @@ class _SendMessagesState extends State<SendMessages> {
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      floatingActionButton: FloatingActionButton(
-        onPressed: () {
-          if (biberCheckBox ||
-              woelfeCheckBox ||
-              meitliCheckBox ||
-              buebeCheckBox) {
-            setState(() {
-              if (_formKey.currentState.validate()) {
-                Map<String, dynamic> data = {
-                  'message': {
-                    'title': titleController.text,
-                    'body': inhaltController.text,
-                    'sender': moreaFire.getPfandiName,
-                    'read': List<String>(),
+    if (!this.initDone) {
+      return Container();
+    } else {
+      return Scaffold(
+        floatingActionButton: FloatingActionButton(
+          onPressed: () {
+            if (groupCheckbox.containsValue(true)) {
+              setState(() {
+                if (_formKey.currentState.validate()) {
+                  groupCheckbox.forEach((key, value) {
+                    if (value) {
+                      this.receivers.add(key);
+                    }
+                  });
+                  Map<String, dynamic> data = {
+                    'message': {
+                      'title': titleController.text,
+                      'body': inhaltController.text,
+                      'sender': moreaFire.getPfandiName,
+                      'read': <String>[],
+                      'receivers': receivers,
+                    },
                     'receivers': receivers,
-                  },
-                  'receivers': receivers,
-                  'snippet': vorschauController.text,
-                  'title': titleController.text
-                };
-                moreaFire.uploadMessage(data);
-                print('Successful: $data');
-                Navigator.of(context).pop();
-              }
-            });
-          } else {
-            print('error');
-            _showDialog(context);
-          }
-        },
-        child: Icon(Icons.send),
-        backgroundColor: MoreaColors.violett,
-      ),
-      appBar: AppBar(
-        title: Text('Nachricht Senden'),
-      ),
-      body: SingleChildScrollView(
-          padding: EdgeInsets.all(20),
-          child: Form(
-            key: _formKey,
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: <Widget>[
-                Text(
-                  'Empfänger',
-                  style: MoreaTextStyle.lable,
-                ),
-                Padding(
-                  padding: EdgeInsets.only(bottom: 15),
-                ),
-                CheckboxListTile(
-                  value: biberCheckBox,
-                  onChanged: (bool val) {
-                    setState(() {
-                      if (val) {
-                        receivers.add(midatanamebiber);
+                    'snippet': vorschauController.text,
+                    'title': titleController.text
+                  };
+                  moreaFire.uploadMessage(data);
+                  print('Successful: $data');
+                  Navigator.of(context).pop();
+                }
+              });
+            } else {
+              print('error');
+              _showDialog(context);
+            }
+          },
+          child: Icon(Icons.send),
+          backgroundColor: MoreaColors.violett,
+        ),
+        appBar: AppBar(
+          title: Text('Nachricht Senden'),
+        ),
+        body: SingleChildScrollView(
+            padding: EdgeInsets.all(20),
+            child: Form(
+              key: _formKey,
+              child: Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: <Widget>[
+                  Text(
+                    'Empfänger',
+                    style: MoreaTextStyle.lable,
+                  ),
+                  Padding(
+                    padding: EdgeInsets.only(bottom: 15),
+                  ),
+                  ListView(
+                      shrinkWrap: true,
+                      physics: NeverScrollableScrollPhysics(),
+                      children: subgroups.map((Map<dynamic, dynamic> group) {
+                        return CheckboxListTile(
+                          title: Text(
+                            group[groupMapgroupNickName],
+                            style: MoreaTextStyle.normal,
+                          ),
+                          value: groupCheckbox[group["groupID"]],
+                          onChanged: (bool value) {
+                            setState(() {
+                              groupCheckbox[group["groupID"]] = value;
+                            });
+                          },
+                        );
+                      }).toList()),
+                  Padding(
+                    padding: EdgeInsets.only(bottom: 20),
+                  ),
+                  Text(
+                    'Betreff',
+                    style: MoreaTextStyle.lable,
+                  ),
+                  Padding(
+                    padding: EdgeInsets.only(bottom: 15),
+                  ),
+                  TextFormField(
+                    focusNode: titleFocus,
+                    style: MoreaTextStyle.textField,
+                    controller: titleController,
+                    decoration: InputDecoration(
+                      helperText: 'Bspw: Fama',
+                      border: OutlineInputBorder(),
+                    ),
+                    onSaved: (newValue) {
+                      setState(() {
+                        titleController.text = newValue;
+                      });
+                    },
+                    onEditingComplete: () {
+                      FocusScope.of(context).requestFocus(vorschauFocus);
+                    },
+                    validator: (value) {
+                      if (value.isEmpty) {
+                        return 'Bitte nicht leer lassen';
                       } else {
-                        receivers.remove(midatanamebiber);
+                        return null;
                       }
-                      biberCheckBox = val;
-                    });
-                  },
-                  title: Text(
-                    'Biber',
-                    style: MoreaTextStyle.normal,
+                    },
                   ),
-                  controlAffinity: ListTileControlAffinity.platform,
-                ),
-                CheckboxListTile(
-                  value: woelfeCheckBox,
-                  onChanged: (bool val) {
-                    setState(() {
-                      if (val) {
-                        receivers.add(midatanamewoelf);
+                  Padding(
+                    padding: EdgeInsets.only(bottom: 20),
+                  ),
+                  Text(
+                    'Vorschau',
+                    style: MoreaTextStyle.lable,
+                  ),
+                  Padding(
+                    padding: EdgeInsets.only(bottom: 15),
+                  ),
+                  TextFormField(
+                    keyboardType: TextInputType.text,
+                    focusNode: vorschauFocus,
+                    controller: vorschauController,
+                    style: MoreaTextStyle.normal,
+                    minLines: 2,
+                    maxLines: 4,
+                    decoration: InputDecoration(
+                      helperText:
+                          'Dieser Text wird in der Benachrichtigung angezeigt.',
+                      border: OutlineInputBorder(),
+                    ),
+                    onSaved: (newValue) {
+                      setState(() {
+                        vorschauController.text = newValue;
+                      });
+                    },
+                    onEditingComplete: () {
+                      FocusScope.of(context).requestFocus(inhaltFocus);
+                    },
+                    validator: (value) {
+                      if (value.isEmpty) {
+                        return 'Bitte nicht leer lassen';
                       } else {
-                        receivers.remove(midatanamewoelf);
+                        return null;
                       }
-                      woelfeCheckBox = val;
-                    });
-                  },
-                  title: Text(
-                    'Wölfe',
-                    style: MoreaTextStyle.normal,
+                    },
                   ),
-                  controlAffinity: ListTileControlAffinity.platform,
-                ),
-                CheckboxListTile(
-                  value: meitliCheckBox,
-                  onChanged: (bool val) {
-                    setState(() {
-                      if (val) {
-                        receivers.add(midatanamemeitli);
+                  Padding(
+                    padding: EdgeInsets.only(bottom: 20),
+                  ),
+                  Text(
+                    'Inhalt',
+                    style: MoreaTextStyle.lable,
+                  ),
+                  Padding(
+                    padding: EdgeInsets.only(bottom: 15),
+                  ),
+                  TextFormField(
+                    focusNode: inhaltFocus,
+                    controller: inhaltController,
+                    style: MoreaTextStyle.normal,
+                    minLines: 5,
+                    maxLines: 1000,
+                    decoration: InputDecoration(
+                      labelText: 'Nachricht',
+                      alignLabelWithHint: true,
+                      helperText:
+                          'Dieser Text wird in der App beim Öffnen der Nachricht angezeigt.',
+                      border: OutlineInputBorder(),
+                    ),
+                    onSaved: (newValue) {
+                      setState(() {
+                        inhaltController.text = newValue;
+                      });
+                    },
+                    validator: (value) {
+                      if (value.isEmpty) {
+                        return 'Bitte nicht leer lassen';
                       } else {
-                        receivers.remove(midatanamemeitli);
+                        return null;
                       }
-                      meitliCheckBox = val;
-                    });
-                  },
-                  title: Text(
-                    'Meitli',
-                    style: MoreaTextStyle.normal,
+                    },
                   ),
-                  controlAffinity: ListTileControlAffinity.platform,
-                ),
-                CheckboxListTile(
-                  value: buebeCheckBox,
-                  onChanged: (bool val) {
-                    setState(() {
-                      if (val) {
-                        receivers.add(midatanamebuebe);
-                      } else {
-                        receivers.remove(midatanamebuebe);
-                      }
-                      buebeCheckBox = val;
-                    });
-                  },
-                  title: Text(
-                    'Buebe',
-                    style: MoreaTextStyle.normal,
-                  ),
-                  controlAffinity: ListTileControlAffinity.platform,
-                ),
-                Padding(
-                  padding: EdgeInsets.only(bottom: 20),
-                ),
-                Text(
-                  'Betreff',
-                  style: MoreaTextStyle.lable,
-                ),
-                Padding(
-                  padding: EdgeInsets.only(bottom: 15),
-                ),
-                TextFormField(
-                  focusNode: titleFocus,
-                  style: MoreaTextStyle.textField,
-                  controller: titleController,
-                  decoration: InputDecoration(
-                    helperText: 'Bspw: Fama',
-                    border: OutlineInputBorder(),
-                  ),
-                  onSaved: (newValue) {
-                    setState(() {
-                      titleController.text = newValue;
-                    });
-                  },
-                  onEditingComplete: () {
-                    FocusScope.of(context).requestFocus(vorschauFocus);
-                  },
-                  validator: (value) {
-                    if (value.isEmpty) {
-                      return 'Bitte nicht leer lassen';
-                    } else {
-                      return null;
-                    }
-                  },
-                ),
-                Padding(
-                  padding: EdgeInsets.only(bottom: 20),
-                ),
-                Text(
-                  'Vorschau',
-                  style: MoreaTextStyle.lable,
-                ),
-                Padding(
-                  padding: EdgeInsets.only(bottom: 15),
-                ),
-                TextFormField(
-                  keyboardType: TextInputType.text,
-                  focusNode: vorschauFocus,
-                  controller: vorschauController,
-                  style: MoreaTextStyle.normal,
-                  minLines: 2,
-                  maxLines: 4,
-                  decoration: InputDecoration(
-                    helperText:
-                        'Dieser Text wird in der Benachrichtigung angezeigt.',
-                    border: OutlineInputBorder(),
-                  ),
-                  onSaved: (newValue) {
-                    setState(() {
-                      vorschauController.text = newValue;
-                    });
-                  },
-                  onEditingComplete: () {
-                    FocusScope.of(context).requestFocus(inhaltFocus);
-                  },
-                  validator: (value) {
-                    if (value.isEmpty) {
-                      return 'Bitte nicht leer lassen';
-                    } else {
-                      return null;
-                    }
-                  },
-                ),
-                Padding(
-                  padding: EdgeInsets.only(bottom: 20),
-                ),
-                Text(
-                  'Inhalt',
-                  style: MoreaTextStyle.lable,
-                ),
-                Padding(
-                  padding: EdgeInsets.only(bottom: 15),
-                ),
-                TextFormField(
-                  focusNode: inhaltFocus,
-                  controller: inhaltController,
-                  style: MoreaTextStyle.normal,
-                  minLines: 5,
-                  maxLines: 1000,
-                  decoration: InputDecoration(
-                    labelText: 'Nachricht',
-                    alignLabelWithHint: true,
-                    helperText:
-                        'Dieser Text wird in der App beim Öffnen der Nachricht angezeigt.',
-                    border: OutlineInputBorder(),
-                  ),
-                  onSaved: (newValue) {
-                    setState(() {
-                      inhaltController.text = newValue;
-                    });
-                  },
-                  validator: (value) {
-                    if (value.isEmpty) {
-                      return 'Bitte nicht leer lassen';
-                    } else {
-                      return null;
-                    }
-                  },
-                ),
-              ],
-            ),
-          )),
-    );
+                  Padding(
+                    padding: EdgeInsets.only(bottom: 60),
+                  )
+                ],
+              ),
+            )),
+      );
+    }
   }
 }
