@@ -1,18 +1,23 @@
+import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
 import 'package:morea/Widgets/standart/moreaTextStyle.dart';
+import 'package:morea/services/morea_firestore.dart';
 
 import '../../morealayout.dart';
 
 class SingleMessagePage extends StatelessWidget {
-  SingleMessagePage(this.message);
+  SingleMessagePage(this.message, this.moreaFire, this.uid);
 
-  final message;
+  final DocumentSnapshot message;
+  final MoreaFirebase moreaFire;
+  final String uid;
 
   @override
   Widget build(BuildContext context) {
+    this.setMessageRead();
     return Scaffold(
       appBar: AppBar(
-        title: Text(message.data['title']),
+        title: Text(message.data()['title']),
       ),
       body: MoreaBackgroundContainer(
         child: SingleChildScrollView(
@@ -24,7 +29,7 @@ class SingleMessagePage extends StatelessWidget {
                 crossAxisAlignment: CrossAxisAlignment.stretch,
                 children: <Widget>[
                   Text(
-                    'Von: ' + message.data['sender'],
+                    'Von: ' + message.data()['sender'],
                     style: MoreaTextStyle.sender,
                   ),
                   MoreaDivider(),
@@ -32,7 +37,7 @@ class SingleMessagePage extends StatelessWidget {
                     padding: EdgeInsets.all(10),
                   ),
                   Text(
-                    message.data['title'],
+                    message.data()['title'],
                     textAlign: TextAlign.center,
                     style: MoreaTextStyle.title,
                   ),
@@ -40,7 +45,7 @@ class SingleMessagePage extends StatelessWidget {
                     padding: EdgeInsets.all(10),
                   ),
                   Text(
-                    message.data['body'],
+                    message.data()['body'],
                     style: MoreaTextStyle.normal,
                   ),
                 ],
@@ -50,5 +55,10 @@ class SingleMessagePage extends StatelessWidget {
         ),
       ),
     );
+  }
+  Future<void> setMessageRead() async {
+    if (!(message['read'].contains(this.uid))){
+      await moreaFire.setMessageRead(uid, message.id);
+    }
   }
 }
