@@ -11,9 +11,9 @@ import 'package:morea/Widgets/standart/info.dart';
 import 'package:morea/services/crud.dart';
 
 class ViewUserProfilePage extends StatefulWidget {
-  ViewUserProfilePage(this.userData, this.moreaFire, this.crud0);
+  ViewUserProfilePage(this.uID, this.moreaFire, this.crud0);
 
-  final Future<DocumentSnapshot> userData;
+  final String uID;
   final CrudMedthods crud0;
   final MoreaFirebase moreaFire;
 
@@ -43,39 +43,43 @@ class _ViewUserProfilePageState extends State<ViewUserProfilePage>
   @override
   Widget build(BuildContext context) {
     return Container(
-      child: FutureBuilder(
-        future: widget.userData,
-        builder:
-            (BuildContext context, AsyncSnapshot<DocumentSnapshot> aSProfile) {
-          if (!aSProfile.hasData) return moreaLoading.loading();
-          profile = aSProfile.data.data;
-          print(profile);
-
-          return Container(
-              child: Scaffold(
-                  appBar: AppBar(
-                    title: Text(profile['Vorname'].toString()),
-                  ),
-                  body: MoreaBackgroundContainer(
-                      child: SingleChildScrollView(
-                    child: MoreaShadowContainer(
-                      child: viewprofile(),
-                    ),
-                  )),
-                  floatingActionButton: new FloatingActionButton(
-                      elevation: 1.0,
-                      child: new Icon(Icons.edit),
-                      backgroundColor: Color(0xff7a62ff),
-                      onPressed: () => Navigator.of(context).push(
-                          new MaterialPageRoute(
-                              builder: (BuildContext context) =>
-                                  new EditUserProfilePage(
-                                      profile: profile,
-                                      moreaFire: widget.moreaFire,
-                                      crud0: widget.crud0))))));
-        },
-      ),
-    );
+        child: Container(
+            child: FutureBuilder<DocumentSnapshot>(
+                future: widget.crud0.getDocument(pathUser, widget.uID),
+                builder: (context, snapshot) {
+                  if (snapshot.connectionState == ConnectionState.waiting) {
+                    return moreaLoading.loading();
+                  } else if (!snapshot.hasData) {
+                    return Scaffold(
+                      appBar: AppBar(
+                        title: Text('Profil nicht vorhanden'),
+                      ),
+                    );
+                  } else {
+                    this.profile = snapshot.data.data();
+                    return Scaffold(
+                        appBar: AppBar(
+                          title: Text(this.profile[userMapVorName]),
+                        ),
+                        body: MoreaBackgroundContainer(
+                            child: SingleChildScrollView(
+                          child: MoreaShadowContainer(
+                            child: viewprofile(),
+                          ),
+                        )),
+                        floatingActionButton: new FloatingActionButton(
+                            elevation: 1.0,
+                            child: new Icon(Icons.edit),
+                            backgroundColor: Color(0xff7a62ff),
+                            onPressed: () => Navigator.of(context).push(
+                                new MaterialPageRoute(
+                                    builder: (BuildContext context) =>
+                                        new EditUserProfilePage(
+                                            profile: profile,
+                                            moreaFire: widget.moreaFire,
+                                            crud0: widget.crud0)))));
+                  }
+                })));
   }
 
   Widget viewprofile() {
@@ -195,7 +199,7 @@ class _ViewUserProfilePageState extends State<ViewUserProfilePage>
   }
 
   List<Widget> parentWidget() {
-    List<Widget> elternWidget = new List();
+    List<Widget> elternWidget = [];
     for (Future<DocumentSnapshot> dSParent in getParentMap()) {
       elternWidget.add(Container(
         child: FutureBuilder(
@@ -205,7 +209,7 @@ class _ViewUserProfilePageState extends State<ViewUserProfilePage>
             if (!aSParent.hasData)
               return simpleMoreaLoadingIndicator();
             else
-              return displayEltern(aSParent.data.data);
+              return displayEltern(aSParent.data.data());
           },
         ),
       ));
@@ -349,9 +353,9 @@ class _ViewUserProfilePageState extends State<ViewUserProfilePage>
   }
 
   List<Future<DocumentSnapshot>> getParentMap() {
-    List<Future<DocumentSnapshot>> elternMap = new List();
+    List<Future<DocumentSnapshot>> elternMap = [];
     Map<String, String> elt = Map<String, String>.from(profile[userMapEltern]);
-    List<String> elternUID = new List();
+    List<String> elternUID = [];
     elternUID.addAll(elt.values);
     elternUID.forEach(
         (uid) => elternMap.add(widget.moreaFire.getUserInformation(uid)));
