@@ -15,7 +15,7 @@ class Profile extends StatefulWidget {
   final auth;
   final MoreaFirebase moreaFire;
   final Map<String, Function> navigationMap;
-  final Firestore firestore;
+  final FirebaseFirestore firestore;
 
   Profile(
       {@required this.auth,
@@ -285,7 +285,7 @@ class _ProfileState extends State<Profile> {
 
   BottomAppBar _bottomAppBarBuilder() {
     if (widget.moreaFire.getPos == "Leiter") {
-      return moreaLeiterBottomAppBar(widget.navigationMap, 'Ändern');
+      return moreaLeiterBottomAppBar(widget.navigationMap, 'Ändern', MoreaBottomAppBarActivePage.profile);
     } else {
       return moreaChildBottomAppBar(widget.navigationMap);
     }
