@@ -1,12 +1,13 @@
+import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
 import 'package:http/http.dart' as http;
 import 'package:morea/Widgets/animated/MoreaLoading.dart';
 import 'package:morea/Widgets/standart/moreaTextStyle.dart';
+import 'package:morea/morea_strings.dart';
 import 'dart:convert';
 import 'package:morea/services/morea_firestore.dart';
 import 'package:morea/morealayout.dart';
 import 'package:intl/intl.dart';
-import 'package:morea/services/utilities/MiData.dart';
 import 'change_teleblitz_abtreten.dart';
 import 'change_teleblitz_antreten.dart';
 import 'change_teleblitz_bemerkung.dart';
@@ -15,7 +16,7 @@ import 'change_teleblitz_mitnehmen.dart';
 import 'change_teleblitz_sender.dart';
 
 class ChangeTeleblitz extends StatefulWidget {
-  final String stufe;
+  final Map<String, dynamic> stufe;
   final String formType;
   final MoreaFirebase moreaFire;
 
@@ -29,12 +30,13 @@ enum FormType { keineAktivitaet, ferien, normal }
 
 class _ChangeTeleblitzState extends State<ChangeTeleblitz>
     with TickerProviderStateMixin {
-  String stufe;
   FormType formType;
   MoreaFirebase moreaFire;
 
   //Variabeln vom Teleblitz
   String name,
+      startTime,
+      endTime,
       datum,
       antreten,
       mapAntreten,
@@ -45,14 +47,16 @@ class _ChangeTeleblitzState extends State<ChangeTeleblitz>
       grund,
       endeFerien,
       id,
-      slug;
-
-  List<String> mitnehmen;
+      slug,
+      rawDate;
+  Map<String, dynamic> stufe;
+  List<dynamic> mitnehmen;
   bool keineAktivitaet, ferien;
   bool archived = false;
   bool draft = false;
   var oldTeleblitz;
   MoreaLoading moreaLoading;
+  DocumentSnapshot groupDoc;
 
   TeleblitzManager teleblitzManager;
 
@@ -93,8 +97,7 @@ class _ChangeTeleblitzState extends State<ChangeTeleblitz>
               appBar: AppBar(
                 title: Text('Loading...'),
               ),
-              body: MoreaBackgroundContainer(
-                  child: moreaLoading.loading()));
+              body: MoreaBackgroundContainer(child: moreaLoading.loading()));
         } else if (snapshot.connectionState == ConnectionState.done) {
           switch (formType) {
             case FormType.keineAktivitaet:
@@ -134,8 +137,8 @@ class _ChangeTeleblitzState extends State<ChangeTeleblitz>
                                 ),
                                 onTap: () => _selectDatum(context),
                                 trailing: Icon(Icons.arrow_forward_ios),
-                                contentPadding:
-                                    EdgeInsets.only(right: 15, left: 15, bottom: 5),
+                                contentPadding: EdgeInsets.only(
+                                    right: 15, left: 15, bottom: 5),
                               ),
                               Padding(
                                   padding: const EdgeInsets.symmetric(
@@ -148,7 +151,7 @@ class _ChangeTeleblitzState extends State<ChangeTeleblitz>
                                 title: Text('Grund des Ausfalls',
                                     style: MoreaTextStyle.lable),
                                 subtitle: Padding(
-                                  padding: const EdgeInsets.only(top:10.0),
+                                  padding: const EdgeInsets.only(top: 10.0),
                                   child: Text(
                                     this.grund,
                                     style: MoreaTextStyle.subtitle,
@@ -162,7 +165,7 @@ class _ChangeTeleblitzState extends State<ChangeTeleblitz>
                                 },
                                 trailing: Icon(Icons.arrow_forward_ios),
                                 contentPadding: EdgeInsets.only(
-                                    right: 15, left: 15, bottom: 15, top:5),
+                                    right: 15, left: 15, bottom: 15, top: 5),
                               ),
                             ],
                           ),
@@ -259,8 +262,8 @@ class _ChangeTeleblitzState extends State<ChangeTeleblitz>
                                 ),
                                 onTap: () => _selectDatum(context),
                                 trailing: Icon(Icons.arrow_forward_ios),
-                                contentPadding:
-                                    EdgeInsets.only(left: 15, right: 15, top: 0, bottom: 5),
+                                contentPadding: EdgeInsets.only(
+                                    left: 15, right: 15, top: 0, bottom: 5),
                               ),
                               Padding(
                                   padding: const EdgeInsets.symmetric(
@@ -288,8 +291,8 @@ class _ChangeTeleblitzState extends State<ChangeTeleblitz>
                                               this.setBeginn)));
                                 },
                                 trailing: Icon(Icons.arrow_forward_ios),
-                                contentPadding:
-                                    EdgeInsets.only(right: 15, left: 15, top: 5, bottom: 5),
+                                contentPadding: EdgeInsets.only(
+                                    right: 15, left: 15, top: 5, bottom: 5),
                               ),
                               Padding(
                                   padding: const EdgeInsets.symmetric(
@@ -302,7 +305,7 @@ class _ChangeTeleblitzState extends State<ChangeTeleblitz>
                                 title: Text('Schluss',
                                     style: MoreaTextStyle.lable),
                                 subtitle: Padding(
-                                  padding: const EdgeInsets.only(top:10.0),
+                                  padding: const EdgeInsets.only(top: 10.0),
                                   child: Text(
                                     abtreten,
                                     style: MoreaTextStyle.subtitle,
@@ -315,8 +318,8 @@ class _ChangeTeleblitzState extends State<ChangeTeleblitz>
                                               this.mapAbtreten, this.setEnde)));
                                 },
                                 trailing: Icon(Icons.arrow_forward_ios),
-                                contentPadding:
-                                    EdgeInsets.only(right: 15, left: 15, top: 5, bottom: 5),
+                                contentPadding: EdgeInsets.only(
+                                    right: 15, left: 15, top: 5, bottom: 5),
                               ),
                               Padding(
                                   padding: const EdgeInsets.symmetric(
@@ -349,8 +352,8 @@ class _ChangeTeleblitzState extends State<ChangeTeleblitz>
                                               this.setMitnehmen)));
                                 },
                                 trailing: Icon(Icons.arrow_forward_ios),
-                                contentPadding:
-                                    EdgeInsets.only(right: 15, left: 15, top: 10, bottom: 10),
+                                contentPadding: EdgeInsets.only(
+                                    right: 15, left: 15, top: 10, bottom: 10),
                               ),
                               Padding(
                                   padding: const EdgeInsets.symmetric(
@@ -376,8 +379,8 @@ class _ChangeTeleblitzState extends State<ChangeTeleblitz>
                                               this.setBemerkung)));
                                 },
                                 trailing: Icon(Icons.arrow_forward_ios),
-                                contentPadding:
-                                    EdgeInsets.only(right: 15, left: 15, top:5, bottom:5),
+                                contentPadding: EdgeInsets.only(
+                                    right: 15, left: 15, top: 5, bottom: 5),
                               ),
                               Padding(
                                   padding: const EdgeInsets.symmetric(
@@ -423,7 +426,10 @@ class _ChangeTeleblitzState extends State<ChangeTeleblitz>
   }
 
   Future<Map> downloadTeleblitz() async {
-    var infos = await teleblitzManager.downloadTeleblitz(this.stufe);
+    this.groupDoc =
+        await widget.moreaFire.crud0.getDocument(pathGroups, stufe['groupID']);
+    var infos =
+        await teleblitzManager.downloadTeleblitz(this.stufe, this.groupDoc);
     this.name = infos['name'];
     this.datum = infos['datum'];
     antreten = infos['antreten'];
@@ -435,8 +441,6 @@ class _ChangeTeleblitzState extends State<ChangeTeleblitz>
     sender = infos['name-des-senders'];
     grund = infos['grund'];
     endeFerien = infos['ende-ferien'];
-    id = infos['_id'];
-    slug = infos['slug'];
     switch (widget.formType) {
       case 'keineAktivitaet':
         keineAktivitaet = true;
@@ -455,29 +459,61 @@ class _ChangeTeleblitzState extends State<ChangeTeleblitz>
   }
 
   void uploadTeleblitz() async {
-    Map<String, dynamic> newTeleblitz = {
-      'name': this.name,
-      'datum': this.datum,
-      'antreten': this.antreten,
-      'google-map': this.mapAntreten,
-      'abtreten': this.abtreten,
-      'map-abtreten': this.mapAbtreten,
-      'mitnehmen-test': this.mitnehmen,
-      'bemerkung': this.bemerkung,
-      'name-des-senders': this.sender,
-      'grund': this.grund,
-      'ende-ferien': this.endeFerien,
-      'slug': this.slug,
-      'ferien': this.ferien,
-      'keine-aktivitat': this.keineAktivitaet,
-      '_draft': false,
-      '_archived': false,
-      'groupIDs': [convWebflowtoMiData(stufe)],
-    };
-    teleblitzManager.uploadTeleblitz(newTeleblitz, this.id);
-    await widget.moreaFire
-        .uploadteleblitz(convWebflowtoMiData(stufe), newTeleblitz);
-    Navigator.of(context).popUntil(ModalRoute.withName('/'));
+    if (this.datum != 'Bitte wählen') {
+      Map<String, dynamic> newTeleblitz = {
+        'name': this.name,
+        'datum': this.datum,
+        'antreten': this.antreten,
+        'google-map': this.mapAntreten,
+        'abtreten': this.abtreten,
+        'map-abtreten': this.mapAbtreten,
+        'mitnehmen-test': this.mitnehmen,
+        'bemerkung': this.bemerkung,
+        'name-des-senders': this.sender,
+        'grund': this.grund,
+        'ende-ferien': this.endeFerien,
+        'ferien': this.ferien,
+        'keine-aktivitat': this.keineAktivitaet,
+        '_draft': false,
+        '_archived': false,
+        'groupIDs': [stufe['groupID']],
+        'EventType': 'Teleblitz',
+        mapTimestamp: DateTime.now().toIso8601String()
+      };
+      int hours = int.parse(this.antreten.substring(0, 2));
+      int minutes = int.parse(this.antreten.substring(3, 5));
+      if (this.groupDoc.get(groupMapGroupOption)['webflowCMSID'] != null) {
+        teleblitzManager.uploadTeleblitz(newTeleblitz,
+            this.groupDoc.get(groupMapGroupOption)['webflowCMSID']);
+      }
+      DateTime start =
+          (DateFormat("dd.MM.yyyy").parse(this.datum.split(', ')[1]))
+              .add(Duration(hours: hours, minutes: minutes));
+      hours = int.parse(this.abtreten.substring(0, 2));
+      minutes = int.parse(this.abtreten.substring(3, 5));
+      DateTime end = (DateFormat("dd.MM.yyyy").parse(this.datum.split(', ')[1]))
+          .add(Duration(hours: hours, minutes: minutes));
+      if (groupDoc.get('groupOption')['teleblitzID'] != null) {
+        await widget.moreaFire.createTeleblitz([stufe['groupID']],
+            start.toIso8601String(), end.toIso8601String(), newTeleblitz,
+            eventID: groupDoc.get('groupOption')['teleblitzID']);
+      } else {
+        await widget.moreaFire.createTeleblitz([stufe['groupID']],
+            start.toIso8601String(), end.toIso8601String(), newTeleblitz);
+      }
+
+      Navigator.of(context).popUntil(ModalRoute.withName('/'));
+    } else {
+      showDialog(
+          context: context,
+          builder: (context) {
+            return AlertDialog(
+                content: Text(
+              'Bitte Datum wählen',
+              style: MoreaTextStyle.warningButton,
+            ));
+          });
+    }
   }
 
   Future<Null> _selectDatum(BuildContext context) async {
@@ -489,6 +525,7 @@ class _ChangeTeleblitzState extends State<ChangeTeleblitz>
         lastDate: now.add(Duration(days: 9999)));
     if (picked != null) {
       setState(() {
+        this.rawDate = picked.toIso8601String();
         this.datum = DateFormat('EEEE, dd.MM.yyy', 'de').format(picked);
       });
     }
@@ -511,11 +548,13 @@ class _ChangeTeleblitzState extends State<ChangeTeleblitz>
   void setBeginn(String ort, String zeit, String map) {
     this.antreten = zeit + ' Uhr, ' + ort;
     this.mapAntreten = map;
+    this.startTime = zeit;
   }
 
   void setEnde(String ort, String zeit, String map) {
     this.abtreten = zeit + ' Uhr, ' + ort;
     this.mapAbtreten = map;
+    this.endTime = zeit;
   }
 
   void setMitnehmen(List<String> mitnehmen) {
@@ -541,81 +580,81 @@ class _ChangeTeleblitzState extends State<ChangeTeleblitz>
 
 class TeleblitzManager {
   MoreaFirebase moreaFirebase;
+  DocumentSnapshot groupDoc;
 
   TeleblitzManager(MoreaFirebase moreaFirebase) {
     this.moreaFirebase = moreaFirebase;
   }
 
-  Future<Map> downloadTeleblitz(String stufe) async {
-    String apiKey = await moreaFirebase.getWebflowApiKey();
-    var jsonDecode;
-    var jsonString;
-    jsonString = await http.get(
-        "https://api.webflow.com/collections/5be4a9a6dbcc0a24d7cb0ee9/items?api_version=1.0.0&access_token=$apiKey");
-    jsonDecode = json.decode(jsonString.body);
-    Map infos;
-    for (var u in jsonDecode['items']) {
-      if (u["name"] == stufe) {
-        infos = u;
-      }
-    }
-    String newDate = _formatDate(infos['ende-ferien']);
-    infos['ende-ferien'] = newDate;
-    infos['mitnehmen-test'] = _formatMitnehmen(infos['mitnehmen-test']);
-    return infos;
-  }
-
-  String _formatDate(String date) {
-    if (date != '') {
-      String rawDate = date.split('T')[0];
-      List<String> dates = rawDate.split('-');
-      String formatedDate = dates[2] + '.' + dates[1] + '.' + dates[0];
-      return formatedDate;
+  Future<Map> downloadTeleblitz(
+      Map<String, dynamic> stufe, DocumentSnapshot groupDoc) async {
+    this.groupDoc = groupDoc;
+    if (groupDoc.get(groupMapGroupOption)['teleblitzID'] != null) {
+      DocumentSnapshot teleblitzDoc = await this
+          .moreaFirebase
+          .crud0
+          .getDocument(
+              pathEvents, groupDoc.get(groupMapGroupOption)['teleblitzID']);
+      return teleblitzDoc.data();
     } else {
-      return '29.10.2019';
+      Map<String, dynamic> teleblitz = {
+        '_archived': false,
+        '_draft': false,
+        'abtreten': '00:00 Uhr, Lokal',
+        'antreten': '00:00 Uhr, Lokal',
+        'bemerkung': 'Bitte definieren',
+        'datum': 'Bitte wählen',
+        'ende-ferien': '22.01.2021',
+        'ferien': false,
+        'google-map': "Bitte definieren",
+        'grund': 'Bitte definieren',
+        'keine-aktivitat': false,
+        'map-abtreten': 'Bitte definieren',
+        'mitnehmen-test': ['Bitte definieren'],
+        'name': stufe['groupNickName'],
+        'name-des-senders': 'Bitte definieren'
+      };
+      return teleblitz;
     }
-  }
-
-  List _formatMitnehmen(String mitnehmen) {
-    List newMitnehmen = mitnehmen
-        .replaceFirst("<ul>", "")
-        .replaceFirst('<' + '/' + 'ul>', "")
-        .replaceAll("</li><li>", ";")
-        .replaceFirst("<li>", "")
-        .replaceFirst("</li>", "")
-        .split(';');
-    return newMitnehmen;
   }
 
   void uploadTeleblitz(Map newTeleblitz, String id) async {
-    String apiKey = await moreaFirebase.getWebflowApiKey();
-    String formatedMitnehmen = '<ul>';
-    for (String u in newTeleblitz['mitnehmen-test']) {
-      formatedMitnehmen = formatedMitnehmen + '<li>' + u + '</li>';
+    if (this.groupDoc.get(groupMapGroupOption)['webflowCMSID'] != null) {
+      String apiKey = await moreaFirebase.getWebflowApiKey();
+      String formatedMitnehmen = '<ul>';
+      for (String u in newTeleblitz['mitnehmen-test']) {
+        formatedMitnehmen = formatedMitnehmen + '<li>' + u + '</li>';
+      }
+      formatedMitnehmen = formatedMitnehmen + '</ul>';
+      newTeleblitz['mitnehmen-test'] = formatedMitnehmen;
+      var result = newTeleblitz['ende-ferien'].split('.');
+      newTeleblitz['ende-ferien'] =
+          result[2] + '-' + result[1] + '-' + result[0] + 'T00:00:00.000Z';
+      newTeleblitz.remove('groupIDs');
+      newTeleblitz.remove('EventType');
+      newTeleblitz.remove('Timestamp');
+      newTeleblitz.remove('_draft');
+      newTeleblitz.remove('_archived');
+      print(id);
+      var jsonMap = {'fields': newTeleblitz};
+      String jsonStr = jsonEncode(jsonMap);
+      print(jsonStr);
+      Map<String, String> header = Map();
+      header["Authorization"] = "Bearer $apiKey";
+      header["accept-version"] = "1.0.0";
+      header["Content-Type"] = "application/json";
+      http
+          .patch(
+        Uri.https(
+            'api.webflow.com',
+            "/collections/5be4a9a6dbcc0a24d7cb0ee9/items/" + id,
+            {'live': 'true'}),
+        headers: header,
+        body: jsonStr,
+      )
+          .then((http.Response result) {
+        print(result.body);
+      });
     }
-    formatedMitnehmen = formatedMitnehmen + '</ul>';
-    newTeleblitz['mitnehmen-test'] = formatedMitnehmen;
-    var result = newTeleblitz['ende-ferien'].split('.');
-    newTeleblitz['ende-ferien'] =
-        result[2] + '-' + result[1] + '-' + result[0] + 'T00:00:00.000Z';
-    newTeleblitz.remove('groupIDs');
-    var jsonMap = {"fields": newTeleblitz};
-    String jsonStr = jsonEncode(jsonMap);
-    Map<String, String> header = Map();
-    header["Authorization"] = "Bearer $apiKey";
-    header["accept-version"] = "1.0.0";
-    header["Content-Type"] = "application/json";
-    http
-        .put(
-      "https://api.webflow.com/collections/5be4a9a6dbcc0a24d7cb0ee9/items/" +
-          id +
-          "?live=true",
-      headers: header,
-      body: jsonStr,
-    )
-        .then((result) {
-      print(result.statusCode);
-      print(result.body);
-    });
   }
 }
