@@ -28,7 +28,7 @@ abstract class BaseTeleblitzFirestore {
 
   Future<bool> eventIDExists(String eventID);
 
-  //Returns true if a given eventID exists in the firestore
+  //Returns true if a given eventID exists in the FirebaseFirestore
 
   Future<void> uploadTelbzAkt(String groupID, Map<String, dynamic> data);
 
@@ -56,7 +56,7 @@ class TeleblizFirestore implements BaseTeleblitzFirestore {
   Stream<Map<String, Map<String, Map<String, dynamic>>>> get getMapofEvents =>
       _mapofEventsController.stream;
 
-  TeleblizFirestore(Firestore firestore, List<String> groupIDs) {
+  TeleblizFirestore(FirebaseFirestore firestore, List<String> groupIDs) {
     crud0 = CrudMedthods(firestore);
     _mapHomeFeedController.addStream(this.streamMapHomeFeed(groupIDs));
     this.streamMapofGroupEvents(groupIDs);
@@ -67,8 +67,7 @@ class TeleblizFirestore implements BaseTeleblitzFirestore {
         crud0.streamDocument(pathGroups, groupID);
     await for (List<String> homeFeed
         in sDhomeFeed.map((DocumentSnapshot dsHomeFeed) {
-      return new List<String>.from(
-          dsHomeFeed.data[groupMapHomeFeed] ?? new List());
+      return new List<String>.from(dsHomeFeed.data()[groupMapHomeFeed] ?? []);
     })) {
       mapHomeFeed[groupID] = homeFeed;
       yield mapHomeFeed;
@@ -84,7 +83,7 @@ class TeleblizFirestore implements BaseTeleblitzFirestore {
 
   Stream<Map<String, dynamic>> steramTelebliz(eventID) async* {
     yield* crud0.streamDocument(pathEvents, eventID).map((dsEvent) {
-      return dsEvent.data;
+      return dsEvent.data();
     });
   }
 
@@ -99,7 +98,7 @@ class TeleblizFirestore implements BaseTeleblitzFirestore {
   Stream<Map<String, Map<String, dynamic>>> steamMapofEvents(
       List<String> eventIDs) {
     Map<String, Map<String, dynamic>> mapOfEvents = new Map();
-    List<Stream<Map<String, Map<String, dynamic>>>> listStream = new List();
+    List<Stream<Map<String, Map<String, dynamic>>>> listStream = [];
     for (String eventID in eventIDs) {
       listStream.add(steamMapofEventshelper(eventID, mapOfEvents));
     }
@@ -117,8 +116,7 @@ class TeleblizFirestore implements BaseTeleblitzFirestore {
   }
 
   void helper(Map<String, List<String>> listHomeFeed) async {
-    List<Stream<Map<String, Map<String, Map<String, dynamic>>>>> list =
-        new List();
+    List<Stream<Map<String, Map<String, Map<String, dynamic>>>>> list = [];
     for (MapEntry<String, List<String>> homeFeed in listHomeFeed.entries) {
       list.add(streamMapofGroupEventsHelper(homeFeed).asBroadcastStream());
     }
@@ -151,7 +149,7 @@ class TeleblizFirestore implements BaseTeleblitzFirestore {
   Future<Map> refeshTelbz(String eventID) async {
     try {
       Map<String, dynamic> tlbz = Map<String, dynamic>.from(
-          (await crud0.getDocument(pathGroups, eventID)).data);
+          (await crud0.getDocument(pathGroups, eventID)).data());
       tlbz["Timestamp"] = DateTime.now();
       this._teleblitze[eventID] = tlbz;
       return tlbz;
@@ -168,7 +166,7 @@ class TeleblizFirestore implements BaseTeleblitzFirestore {
       if (!dSAnmeldung.exists)
         yield "un-initialized";
       else
-        yield dSAnmeldung.data["AnmeldeStatus"];
+        yield dSAnmeldung.data()["AnmeldeStatus"];
     }
   }
 
@@ -177,8 +175,8 @@ class TeleblizFirestore implements BaseTeleblitzFirestore {
     return doc.exists;
   }
 
-  Future<void> uploadTelbzAkt(String groupnr, Map<String, dynamic> data) async {
-    return await crud0.runTransaction(pathGroups, groupnr, data);
+  Future<void> uploadTelbzAkt(String groupID, Map<String, dynamic> data) async {
+    return await crud0.runTransaction(pathGroups, groupID, data);
   }
 
   Future<void> uploadTelbz(String eventID, Map<String, dynamic> data) async {
