@@ -11,13 +11,13 @@ String convWebflowtoMiData(String stufe) {
     case 'NahaniMeitli':
       return '3779';
     case 'DrasonBuebe':
-      return '4013';
+      return 'Op3qqioWs36MCeoWgDYL';
     case 'Wombat (Wölfe)':
       return '3776';
     case 'Nahani (Meitli)':
       return '3779';
     case 'Drason (Buebe)':
-      return '4013';
+      return 'Op3qqioWs36MCeoWgDYL';
     default:
       print("stufe = " + stufe);
       return stufe;
@@ -26,15 +26,22 @@ String convWebflowtoMiData(String stufe) {
 
 String convMiDatatoWebflow(String groupID) {
   switch (groupID) {
-    case '3775':
+    case 'hjZL6PT8t8MCfxf5GOko':
       return 'Biber';
-    case '3776':
-      return 'Wombat (Wölfe)';
-    case '3779':
+    case 'TJNO3BHaPnBZE33anezc':
+      return 'Wölfe 1';
+    case 'VVy3DAlQLmbw9oGuRwvZ':
       return 'Nahani (Meitli)';
     case '4013':
       return 'Drason (Buebe)';
+    case 'Op3qqioWs36MCeoWgDYL':
+      return 'Drason (Buebe)';
+    case 'UisNaofVTp924gvx6Kcf':
+      return 'Wölfe 2';
+    case 'iRHtWXGhstTpMwmceqaF':
+      return 'Pio';
     default:
-      throw "convMiDatatoWebflow Error groupID given was: " + groupID.toString();
+      throw "convMiDatatoWebflow Error groupID given was: " +
+          groupID.toString();
   }
 }
